requirements.txt
pandas
matplotlib
scikit-learn
tensorflow-cpu
opencv-python
numpy